Backup/V3_Peer_Review/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
PyPDF2>=3.0.0 